requirements.txt
boto3
pandas >= 0.23.4
numpy >= 1.15.0
psycopg2-binary
scikit-learn >=0.19.2